movieapp/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
idna==3.4
Pillow==9.2.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
